+++ NEWS.md
@@ -4,6 +4,10 @@
 
 - `pearsons_c()` effect size column name changed to `Pearsons_c` for consistency. 
 
+## New features
+
+- `rbs_to_common_language()` for converting rank-biserial correlation to Probability of superiority.
+
 ## Changes
 
 - `effectsize()` for `BayesFactor` objects returns the same standardized output as for `htest`.